Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d368330-7ead-4276-b2bc-6b6cf65f6ea3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383200176
Change-Id: Ic660c6fc1ddb863028e96664a51c11c76490d20a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com