Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ffbf597d-b7e2-451b-95ce-e0dfd9c1927f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383200240
Change-Id: I7b2e021cd115486b460704490c5b13471612c369
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com