Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e03404e-e8bf-41bd-bbee-577ff8c81295). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383199717
Change-Id: Ib65b4c494d5fe643495a4b1c762a1877705c7d51
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com