Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1ac3dd1-5949-4875-9cf0-c74ca7c100a8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383199953
Change-Id: Ie71f6660d1a78ace7f54a63fff1cfa980395bbb0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com