Revert "Keep _FORTIFY_SOURCE=2 for libvulkan when enabling _FORT..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f5e5fbd-8072-4664-b443-7aecc8956d73). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 383199928
Change-Id: I8d8468c3fa53e020071370eda36a5f644de4be06
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com