Revert "Add Accessor::from_binder"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46e6e9df-23c7-4968-8ddd-070549ce886f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id4993c45b65f97062324948e3ffcfdecef9dcd60
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com