Revert "Use local copy of vm_sockets.h on non-Trusty for portability"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a80ad64-5292-4aa4-90e9-36fb4a93a042). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61f7552ae6490044b51ee3af4fbb01a473664105
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com