dumpstate: Don't assume cgroupfs layout
Dumpstate currently assumes the cgroupfs layout will always be
/uid_x/pid_y, but paths used for v2 cgroups in Android are owned by
libprocessgroup, and change based on build flags such as
PRODUCT_CGROUP_V2_SYS_APP_ISOLATION_ENABLED which is about to be
defaulted to true. So that assumption is not always correct.

Update dumpstate to recurse until it finds cgroups starting with "pid_"
regardless of the depth/level, and regardless of the name of the parent
cgroup.

Bug: 382693152
Change-Id: Iad3f6cb28a3fb7aa33b34e03787f9df9276da392