vkjson: Move redefinition of FALLTHROUGH_INTENDED
The macro FALLTHROUGH_INTENDED is defined in both
vulkan/vkjson/vkjson.h and android-base/macros.h, causing build errors
for any other modules that include both header files.

Move the redefinition to vkjson.cc.

Note that <android-base/macros.h> is not available for the libvkjson_ndk
build, so we cannot use that version of the macro and eliminate the
redefinition entirely.

Bug: 372694741
Test: CQ
Change-Id: I385808571cade7dd0b3eeecd96344da6b137ba30