Revert "Remove the extra parameters in onCommitNotComposited"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47109a0a-36ec-462f-859f-5e24aa032aa8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie9ffa82c9e2367209d6b7a5e5d198314df80bfa5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com