Revert "Cleanup GraphicsEnv::setAngleInfo()"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/837bb21e-5ade-45ed-9b2a-c268b62d4f67). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9499f5f0bb0494b39457ec701776160e8bb7723f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com