trusty: Don't generate errors for new lints in clang/LLVM 19
clang/LLVM 19 triggers a new lint (-Wunused-but-set-variable) when
building Parcel.cpp for Trusty because `BINDER_WITH_KERNEL_IPC` is not
defined. Trusty neededs to roll to clang 19 to match the version of
LLVM used to build Rust 1.82.

The lints are raised because some variables are set but unused whenever
BINDER_WITH_KERNEL_IPC is undefined. Since the lints are not surfacing
actual problems, we silence them by marking the variables in question
as unused. An alternative, less targeted solution would be to not treat
this lint as an error in Trusty builds or any build where
BINDER_WITH_KERNEL_IPC is undefined.

Bug: 390243478
Test: Treehugger
Test: build Trusty with prebuilt clang-r536225
Change-Id: Ib45a21b79ca87f5467b67d0d9179cdfb2e4c0331