Revert "Remove redundant local parameters in Client"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/902d7788-b653-433b-b72e-cd2056bde564). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d4e4cbeebc455bf5c9425b8ba11780bc956642e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com