Revert "Remove redundant local parameters in Client"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca54df86-f5ee-459a-86b2-5e05c7d2e60a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4038c73ff63be4641e86c3a8a496e83040b3baa5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com