Revert "Remove redundant local parameters in Client"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ffda1c3f-9ca3-4d15-8318-2cb6d0c51086). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I649e417cd07b90def14dd86840c94a0233582dee
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com