Revert "Enable searching for mounted public volumes under any user"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/436f025f-3209-40d9-a6a8-da90b1435087). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8956d76c620eb880e3f9cff82bb9ac1e82549375
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com