Revert "RPC Binder: increase transaction size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3fb1a720-d8c7-4caf-8397-3dd27fd365f7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 392992112
Change-Id: I0a0182f62dd9442fc4c68ca1357a475e43cbced9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com