Revert "Enable searching for mounted public volumes under any user"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efdfe5fb-4013-4bf0-81ea-36b44d8dc89b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85f7eccf3d5b30314ae36b9829341beceb6320ce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com