Revert "libgui InputTransferToken: avoid implicit sp<>"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/380b14b4-46a5-4b25-8498-d5f0483b501d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0a1b55a2f670347e13d4c8fd03f58eaebe2680bd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com