Revert "opengl stubs: update for no implicit sp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/380b14b4-46a5-4b25-8498-d5f0483b501d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib1f88713502f4a2f6dad38a4c74e2f63b1408c67
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com