Report input usage session snapshots
There could be delayed or missing data when logging usage
sessions at the start of the next session. There is no way to
query if there is a current active or unrecorded session, or
to force record inactive sessions. As a solution, log snapshots
of input usage sessions every 5 minutes while devices are being used.

Test: Verified that session snapshot events are logged to statsd
when buttons are pressed 5 minutes since the previous action.

Change-Id: Ie20e771a14f15ef59d75b3179c44c073e049f986