Revert "Adding secure hal interfaces to binder rpc visibility"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b91feab5-b64a-47cf-ba2c-56b04624695b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3249b9bc3b280bb77bb656a89ececa8737c589ec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com