Revert "Adding secure hal interfaces to binder rpc visibility"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d5dc844a-b28b-420c-aa37-32d9ad24f57c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I80506667ca6e8c09e9509ffc96103645a35b713e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com