libbinder: allow fd transport in trusty by default
Removes the default protocol version of 0 for trusty servers, which
keeps trusty servers in line with the RPC_WIRE_PROTOCOL_VERSION
(currently 1). This latest version allows parcels to contain FDs. This
should be safe because rpc binder negotiates a protocol version using
the min value of the client and server configurations. Older clients
will be able to downgrade the protocol version (but not transact with
fds in their parcels) and newer clients will see their protocol versions
negotiated down to the server's version on connection.

This change also adds supported file descriptor transport modes to
servers by default. In trusty, tipc allows handles to be transmitted by
default so this is in line with the current security posture.

Bug: 377907450
Test: WIP
Change-Id: I0efbd7781f0fd2bc81981b01ae1b1d0d149b12b6