Revert "Fix the issue of screen flashing when adjusting ambient ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c818a7fe-ac9c-45ee-934c-6eab54c4ea50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0980c7bebbeec1f5b7f9c545474dcb08e6a0925
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com