IServiceManager add checkServiceAccess to delegate sepolicy check
Android Virtualization Framework is in charge of VSOCK communications on
Android devices. Specifically, virtmgr manages these connection for AVF
clients that own VMs. It is forked from the owner and can get the
selinux SID of that owning process.
We want to check that the owning process has access to services that it
is requesting.
So we provide a new checkServiceAccess method that takes the SID of the
owner and checks if it has access to "find" a given service before
virtmgr provides access to it.

Test: atest vm_accessor_test
Bug: 358427181
Change-Id: Ie389af62f1af541a554294d79f72b4c3b65c3e45