Revert "Add meta transactions in fuzzService"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d42f20e0-1528-4dd8-9696-0d429e65a6c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icab926b0003f8a8c41d489646a82cff11349e136
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com