Revert "Remove log if cache is disabled."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/337f79e1-e35b-4c1d-bf37-9c6a7e2069c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ef45a781be4124649959c7f4a85115a24d56930
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com