Drop the buffer from the cache if the buffer is dropped from the transaction
Summary: The initial issue is that the rendering system and the client end up with mismatched cache information. The client thinks a buffer is cached, but the rendering system fails to find the buffer, therefore it's unable to render the new content.

This can happen in the following scenario:
1. An application sets buffer1 into a transaction but doesn't apply it. The buffer is then cached locally in the process.
2. The transaction is sent to another process.
3. The application sets buffer2 into a new transaction but doesn't apply it.
4. The second transaction is sent to the other process.
5. The two transactions are merged, causing the buffer in the second transaction to overwrite the buffer in the first transaction.

This causes the code to invoke a release callback back to the original process so it can release its buffer since it's essentially a drop. However, the client still thinks the buffer is cached, but it will never get cached in the rendering system because the buffer was dropped before being sent.

In this change, it notifies the calling process to also uncache the buffer when releasing, so the next time the buffer is used, it will get cached in both the rendering system and the client.

Test: Successful Build on master branch

Change-Id: Id1b65f9d28112c5f9083e6cc7685548d26d5e16a
Signed-off-by: Pratik Kathalkar <pkathalk@meta.com>