GUI: phase out writeToParcel cacheBuffers
When the app still has transactioncallback,it will not immediately apply
the transaction to SurfaceFlinger. At this time,the app will call
'writeToParcel' to invoke 'cacheBuffer' and add the cache on the Client
Side. But does not call the server-side ClientCache to add the cache,
then in the next frame, when the server tries to get the buffer using
the cache ID, it will report an error.

BUG: 438100090
Test: Manual
1. android device go to split screen mode, one activity in 1/3 region,
another activity in 2/3 region
2. Press and hold the operation button in the middle of the split screen
to adjust the window size and swap the left and right windows, with frequent
operations.
3. run cmd to kill and then start, package com.test run in 2/3 region while true;
do echo 'waiting...';adb shell pkill com.test;sleep 1;adb shell am start -n
com.test/.MainActivity sleep 5;done
4.adb -d shell "monkey --pct-touch 40 --pct-motion 30 --pct-pinchzoom 5
--pct-syskeys 10 --pct-appswitch 10 --pct-flip 5 -s 0 --ignore-crashes
--ignore-timeouts --ignore-native-crashes --throttle 500 -v -v -v 500000"

Change-Id: I890b18c10a2f5fca454dce0bf36fa0452b4f9e00