Fix the issue that inconsistency between SurfaceFlinger property settings and framework-side static variable initialization.
Bug:444122400
Test:Manual

When repeatedly performing the operation of adb reboot recovery to clear all data and reboot, it occasionally occurs that static constants in the WindowManagerService class, obtained via SystemProperties.get, retrieve properties from the initBootProperties function within the asynchronous task init in the SurfaceFlinger class that have not been initialized. This leads to judgment anomalies when using these static constants in WindowManagerService.

Change-Id: I8f7ed9f86f3741ce53a94d59c271f0eb8931d521
Signed-off-by: zhengchuanlong@xiaomi.com <zhengchuanlong@xiaomi.com>