Null Pointer native crash in system_server while calling InputDispatcher::injectInputEvent
A Null Pointer Exception (NPE) occurs in InputDispatcher::injectInputEvent causing the
system_server process to crash. The exact test scenario is random.

The crash is observed while call made to injectInputEvent function which dereferenced
event pointer in a log line. The stack trace pointed out the dereferenced event pointer crash
on trying to access a keyevent which does not have mapping to a label, pointing out that
the label string is null, resulting in crash. So with the fix we are trying to
handle any random keycode event which is not mapped to a label should not result in a
system_server crash

Bug: 465464500

Change-Id: I4788ace6a57a42ecdd8464405161a5cb74f90e43
Signed-off-by: Ayush Garg <agarg9@ford.com>