JankTracker::onJankData add quickpath
In JankTracker::onJankData, certain conditions can be detected in advance. If the conditions are not met, the thread scheduling overhead of BackgroundExecutor.getLowPriorityInstance can be avoided.
Additionally, to prevent a priority inversion between the SurfaceFlinger main thread and the getLowPriorityInstance, the try_lock mechanism is employed. The detection is only performed when the SurfaceFlinger main thread successfully acquires the lock; otherwise, the detection is still carried out within the getLowPriorityInstance.

Change-Id: I29b73f661d16e2db51ca5914c12a99a761ccc758
Signed-off-by: Liangliang Sui <coolsui.coding@gmail.com>